Remove unneeded eip7843 spec.py
