+++ src/ethereum/forks/amsterdam/blocks.py
@@ -114,7 +114,7 @@ class Header:
     [`keccak256`]: ref:ethereum.crypto.hash.keccak256
     [changes]: ref:ethereum.state.State.compute_state_root_and_trie_changes
     [Trie]: ref:ethereum.merkle_patricia_trie.Trie
-    """  # noqa: E501
+    """
 
     transactions_root: Root
     """
@@ -257,6 +257,7 @@ class Header:
     [EIP-7928]: https://eips.ethereum.org/EIPS/eip-7928
     [cbalh]: ref:ethereum.forks.amsterdam.block_access_lists.hash_block_access_list
     """  # noqa: E501
+
     slot_number: U64
     """
     The slot number of this block as provided by the consensus layer.
